+++ src/model/TinyAbomination.java
@@ -15,19 +15,19 @@ public class TinyAbomination {
 	// B929
 	public static double mainHandProcs(Character character) {
 		// TODO: EXPAND
-		return (double) 0;
+		return 0;
 	}
 
 	// B930
 	public static double offHandProcs(Character character) {
 		// TODO: EXPAND
-		return (double) 0;
+		return 0;
 	}
 
 	// B953 * ...
 	private static double mainHandDPS(Character character) {
 		// TODO: EXPAND
-		return ((double) 0)
+		return (0)
 				* ArmorPenetration.mitigation(character,
 						character.mainHand.weapon) * character.bloodFrenzy()
 				* character.hysteria();
@@ -36,7 +36,7 @@ public class TinyAbomination {
 	// B957 * ...
 	private static double offHandDPS(Character character) {
 		// TODO: EXPAND
-		return ((double) 0)
+		return (0)
 				* ArmorPenetration.mitigation(character,
 						character.offHand.weapon) * character.bloodFrenzy()
 				* character.hysteria();
